refactor(scanner): extract task queue and spill records
