+++ src/runtime/scanner.rs
@@ -1,11 +1,15 @@
-use std::collections::VecDeque;
+mod task_queue;
+mod task_spill;
+
 #[cfg(windows)]
 use std::ffi::OsString;
 use std::fs::{self, File};
-use std::io::{self, Read, Seek, SeekFrom, Write};
+use std::io;
 use std::path::{Component, Path, PathBuf};
+use std::sync::Arc;
+#[cfg(all(test, unix))]
+use std::sync::Mutex;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Condvar, Mutex};
 use std::thread;
 
 use cap_primitives::ambient_authority;
@@ -13,23 +17,16 @@ use cap_primitives::fs::{self as cap_fs};
 use crossbeam_channel::Sender;
 use ignore::gitignore::{Gitignore, GitignoreBuilder};
 
+use task_queue::{DirectoryTask, TASK_QUEUE_PER_WORKER, TaskQueue};
+
 use super::worker::{ScannedEntry, WorkerEvent, send_event};
 use crate::model::UnscannedReason;
-use crate::native_path::{EncodedNativePath, NativeIdentity, NativePath, identity_for};
-use crate::temporary_storage::{TemporaryStorage, TemporaryStorageReservation};
+use crate::native_path::{NativeIdentity, identity_for};
+use crate::temporary_storage::TemporaryStorage;
 
 // The owner stages each batch and applies one entry per input poll. Keep batches
 // bounded so backpressure stays small while the map is navigating.
 const BATCH_SIZE: usize = 32;
-
-const TASK_QUEUE_PER_WORKER: usize = 8;
-const MAX_SPILLED_TASK_BYTES: usize = 1024 * 1024;
-const SPILL_LENGTH_BYTES: u64 = 8;
-const TASK_SPILL_COMPACTION_BYTES: u64 = 64 * 1024;
-#[cfg(windows)]
-type TaskSpillFile = tempfile::NamedTempFile;
-#[cfg(not(windows))]
-type TaskSpillFile = File;
 
 #[derive(Clone, Debug)]
 pub struct ScannerOptions {
@@ -40,299 +37,6 @@ pub struct ScannerOptions {
     pub exclusions: Vec<String>,
     pub internal_paths: Vec<PathBuf>,
     pub temporary_storage: TemporaryStorage,
-}
-
-#[derive(Clone)]
-struct DirectoryTask {
-    path: PathBuf,
-    identity: Option<NativeIdentity>,
-}
-
-struct TaskQueue {
-    state: Mutex<QueueState>,
-    ready: Condvar,
-    capacity: usize,
-    root: PathBuf,
-}
-
-struct QueueState {
-    tasks: VecDeque<DirectoryTask>,
-    spill: TaskSpill,
-    pending: usize,
-}
-
-struct TaskSpill {
-    file: TaskSpillFile,
-    next_read: u64,
-    next_write: u64,
-    pending: usize,
-    reservation: TemporaryStorageReservation,
-}
-
-impl TaskSpill {
-    fn new(temporary_storage: &TemporaryStorage) -> io::Result<(Self, Option<PathBuf>)> {
-        #[cfg(windows)]
-        let (file, spill_path) = {
-            let file = tempfile::NamedTempFile::new()?;
-            crate::os::windows::restrict_private_path(file.path(), false)?;
-            crate::os::windows::verify_private_path(file.path(), false)?;
-            let spill_path = Some(file.path().to_path_buf());
-            (file, spill_path)
-        };
-        #[cfg(not(windows))]
-        let (file, spill_path) = (tempfile::tempfile()?, None);
-
-        Ok((
-            Self {
-                file,
-                next_read: 0,
-                next_write: 0,
-                pending: 0,
-                reservation: temporary_storage.reservation(0)?,
-            },
-            spill_path,
-        ))
-    }
-
-    fn push(&mut self, task: DirectoryTask) -> io::Result<()> {
-        let encoded = NativePath::new(task.path).encode();
-        let payload = serde_json::to_vec(&(encoded, task.identity)).map_err(io::Error::other)?;
-        if payload.len() > MAX_SPILLED_TASK_BYTES {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                "scanner task path exceeds the spill record limit",
-            ));
-        }
-        let payload_len = u64::try_from(payload.len()).map_err(|_| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                "scanner task path length does not fit in a spill record",
-            )
-        })?;
-        let next_write = self
-            .next_write
-            .checked_add(SPILL_LENGTH_BYTES)
-            .and_then(|offset| offset.checked_add(payload_len))
-            .ok_or_else(|| io::Error::other("scanner task spill offset overflow"))?;
-        let pending = self
-            .pending
-            .checked_add(1)
-            .ok_or_else(|| io::Error::other("scanner task spill count overflow"))?;
-
-        self.reservation.grow_to(next_write)?;
-        self.file.seek(SeekFrom::Start(self.next_write))?;
-        self.file.write_all(&payload_len.to_le_bytes())?;
-        self.file.write_all(&payload)?;
-        self.next_write = next_write;
-        self.pending = pending;
-        Ok(())
-    }
-
-    fn take(&mut self) -> io::Result<Option<DirectoryTask>> {
-        if self.pending == 0 {
-            return Ok(None);
-        }
-
-        self.file.seek(SeekFrom::Start(self.next_read))?;
-        let mut length = [0_u8; std::mem::size_of::<u64>()];
-        self.file.read_exact(&mut length)?;
-        let payload_len = usize::try_from(u64::from_le_bytes(length)).map_err(|_| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                "scanner task spill record length does not fit in memory",
-            )
-        })?;
-        if payload_len > MAX_SPILLED_TASK_BYTES {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                "scanner task spill record exceeds the configured limit",
-            ));
-        }
-        let mut payload = vec![0_u8; payload_len];
-        self.file.read_exact(&mut payload)?;
-        let (encoded, identity): (EncodedNativePath, Option<NativeIdentity>) =
-            serde_json::from_slice(&payload).map_err(io::Error::other)?;
-        let path = NativePath::decode(&encoded)
-            .map_err(io::Error::other)?
-            .as_path()
-            .to_path_buf();
-        let record_len = u64::try_from(payload_len)
-            .map_err(|_| io::Error::other("scanner task spill record length overflow"))?;
-        self.next_read = self
-            .next_read
-            .checked_add(SPILL_LENGTH_BYTES)
-            .and_then(|offset| offset.checked_add(record_len))
-            .ok_or_else(|| io::Error::other("scanner task spill offset overflow"))?;
-        self.pending = self
-            .pending
-            .checked_sub(1)
-            .ok_or_else(|| io::Error::other("scanner task spill count underflow"))?;
-        if self.pending == 0 {
-            self.next_read = 0;
-            self.next_write = 0;
-            self.truncate(0)?;
-            self.reservation.shrink_to(0);
-        } else {
-            self.compact_after_read()?;
-        }
-        Ok(Some(DirectoryTask { path, identity }))
-    }
-
-    fn truncate(&mut self, len: u64) -> io::Result<()> {
-        #[cfg(windows)]
-        {
-            self.file.as_file().set_len(len)
-        }
-        #[cfg(not(windows))]
-        {
-            self.file.set_len(len)
-        }
-    }
-
-    fn compact_after_read(&mut self) -> io::Result<()> {
-        let remaining = self
-            .next_write
-            .checked_sub(self.next_read)
-            .ok_or_else(|| io::Error::other("scanner task spill offsets are out of order"))?;
-        if remaining == 0 {
-            return Err(io::Error::other(
-                "scanner task spill has queued tasks without record data",
-            ));
-        }
-        let compact_after = self.reservation.bytes().min(TASK_SPILL_COMPACTION_BYTES);
-        if self.next_read < compact_after && self.next_read < remaining {
-            return Ok(());
-        }
-
-        let buffer_len = usize::try_from(remaining.min(TASK_SPILL_COMPACTION_BYTES))
-            .map_err(|_| io::Error::other("scanner task spill compaction buffer is too large"))?;
-        let mut buffer = vec![0_u8; buffer_len];
-        let mut source = self.next_read;
-        let mut destination = 0_u64;
-        let mut bytes_left = remaining;
-        while bytes_left > 0 {
-            let chunk =
-                usize::try_from(bytes_left.min(TASK_SPILL_COMPACTION_BYTES)).map_err(|_| {
-                    io::Error::other("scanner task spill compaction chunk is too large")
-                })?;
-            self.file.seek(SeekFrom::Start(source))?;
-            self.file.read_exact(&mut buffer[..chunk])?;
-            self.file.seek(SeekFrom::Start(destination))?;
-            self.file.write_all(&buffer[..chunk])?;
-            let chunk = u64::try_from(chunk)
-                .map_err(|_| io::Error::other("scanner task spill compaction chunk overflow"))?;
-            source = source
-                .checked_add(chunk)
-                .ok_or_else(|| io::Error::other("scanner task spill offset overflow"))?;
-            destination = destination
-                .checked_add(chunk)
-                .ok_or_else(|| io::Error::other("scanner task spill offset overflow"))?;
-            bytes_left = bytes_left
-                .checked_sub(chunk)
-                .ok_or_else(|| io::Error::other("scanner task spill compaction underflow"))?;
-        }
-        self.truncate(remaining)?;
-        self.reservation.shrink_to(remaining);
-        self.next_read = 0;
-        self.next_write = remaining;
-        Ok(())
-    }
-}
-
-impl TaskQueue {
-    fn new(
-        root: PathBuf,
-        capacity: usize,
-        temporary_storage: &TemporaryStorage,
-    ) -> io::Result<(Self, Option<PathBuf>)> {
-        let (spill, spill_path) = TaskSpill::new(temporary_storage)?;
-        Ok((
-            Self {
-                state: Mutex::new(QueueState {
-                    tasks: VecDeque::from([DirectoryTask {
-                        path: root.clone(),
-                        identity: None,
-                    }]),
-                    spill,
-                    pending: 1,
-                }),
-                ready: Condvar::new(),
-                capacity,
-                root,
-            },
-            spill_path,
-        ))
-    }
-
-    fn take(
-        &self,
-        cancelled: &AtomicBool,
-        failed: &AtomicBool,
-        root_invalid: &AtomicBool,
-    ) -> io::Result<Option<DirectoryTask>> {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        loop {
-            if cancelled.load(Ordering::Acquire)
-                || failed.load(Ordering::Acquire)
-                || root_invalid.load(Ordering::Acquire)
-                || state.pending == 0
-            {
-                return Ok(None);
-            }
-            if let Some(task) = state.tasks.pop_front() {
-                validate_task_path(&self.root, &task.path)?;
-                return Ok(Some(task));
-            }
-            if let Some(task) = state.spill.take()? {
-                validate_task_path(&self.root, &task.path)?;
-                return Ok(Some(task));
-            }
-            state = self
-                .ready
-                .wait(state)
-                .unwrap_or_else(std::sync::PoisonError::into_inner);
-        }
-    }
-
-    fn schedule(&self, task: DirectoryTask) -> io::Result<()> {
-        validate_task_path(&self.root, &task.path)?;
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        let pending = state
-            .pending
-            .checked_add(1)
-            .ok_or_else(|| io::Error::other("scanner task count overflow"))?;
-        if state.tasks.len() < self.capacity {
-            state.tasks.push_back(task);
-        } else {
-            state.spill.push(task)?;
-        }
-        state.pending = pending;
-        self.ready.notify_one();
-        Ok(())
-    }
-
-    fn complete(&self) {
-        let mut state = self
-            .state
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        state.pending = state.pending.saturating_sub(1);
-        self.ready.notify_all();
-    }
-
-    fn cancel(&self) {
-        self.ready.notify_all();
-    }
-}
-
-fn validate_task_path(root: &Path, path: &Path) -> io::Result<()> {
-    task_relative_path(root, path).map(|_| ())
 }
 
 struct Exclusions {
@@ -1335,160 +1039,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn task_queue_spills_overflow_without_growing_the_resident_queue() {
-        let (queue, task_spill_path) =
-            TaskQueue::new(PathBuf::from("/scan-root"), 1, &TemporaryStorage::default())
-                .expect("scanner task spill should be available");
-        #[cfg(windows)]
-        {
-            let task_spill_path = task_spill_path.expect("Windows task spill should be named");
-            assert!(task_spill_path.is_absolute());
-            crate::os::windows::verify_private_path(&task_spill_path, false)
-                .expect("Windows task spill should have a private DACL");
-            let exclusions = Exclusions::new(
-                Path::new("/scan-root"),
-                Vec::new(),
-                vec![task_spill_path.clone()],
-            )
-            .expect("scanner task spill should be an exact internal exclusion");
-            assert_eq!(
-                exclusions.reason(&task_spill_path, false),
-                Some("Excise session state".to_string())
-            );
-        }
-        #[cfg(not(windows))]
-        assert!(task_spill_path.is_none());
-        for index in 0..BATCH_SIZE {
-            queue
-                .schedule(DirectoryTask {
-                    path: PathBuf::from(format!("/scan-root/entry-{index}")),
-                    identity: None,
-                })
-                .expect("scanner task should spill when the resident queue is full");
-        }
-
-        let state = queue
-            .state
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        assert_eq!(state.tasks.len(), 1);
-        assert_eq!(state.spill.pending, BATCH_SIZE);
-    }
-
-    #[test]
-    fn task_queue_enforces_a_total_spill_limit_and_reclaims_drained_records() {
-        let root = PathBuf::from("/scan-root");
-        let task = DirectoryTask {
-            path: root.join("queued"),
-            identity: None,
-        };
-        let payload = serde_json::to_vec(&(
-            NativePath::new(task.path.clone()).encode(),
-            task.identity.clone(),
-        ))
-        .expect("task fixture should serialize");
-        let record_bytes = SPILL_LENGTH_BYTES
-            .checked_add(u64::try_from(payload.len()).expect("task fixture length should fit"))
-            .expect("task fixture record size should fit");
-        let temporary_storage = TemporaryStorage::with_limit_bytes(record_bytes);
-        let (queue, _) = TaskQueue::new(root.clone(), 1, &temporary_storage)
-            .expect("scanner task queue should open");
-
-        queue
-            .schedule(task.clone())
-            .expect("one spill record should fit the session limit");
-        let error = queue
-            .schedule(task)
-            .expect_err("a second spill record must not exceed the session limit");
-        assert_eq!(error.kind(), io::ErrorKind::StorageFull);
-        assert!(error.to_string().contains("--temporary-storage-mib"));
-        {
-            let state = queue
-                .state
-                .lock()
-                .unwrap_or_else(std::sync::PoisonError::into_inner);
-            assert_eq!(state.tasks.len(), 1);
-            assert_eq!(state.spill.pending, 1);
-        }
-        assert_eq!(temporary_storage.used(), record_bytes);
-
-        let cancelled = AtomicBool::new(false);
-        let failed = AtomicBool::new(false);
-        let root_invalid = AtomicBool::new(false);
-        let root_task = queue
-            .take(&cancelled, &failed, &root_invalid)
-            .expect("resident task should be available")
-            .expect("resident task should exist");
-        assert_eq!(root_task.path, root);
-        queue.complete();
-        let queued_task = queue
-            .take(&cancelled, &failed, &root_invalid)
-            .expect("spilled task should be available")
-            .expect("spilled task should exist");
-        assert_eq!(queued_task.path, root.join("queued"));
-        queue.complete();
-        assert_eq!(temporary_storage.used(), 0);
-    }
-
-    #[test]
-    fn task_spill_compacts_consumed_records_before_the_queue_drains() {
-        let root = PathBuf::from("/scan-root");
-        let task = DirectoryTask {
-            path: root.join("queued"),
-            identity: None,
-        };
-        let payload = serde_json::to_vec(&(
-            NativePath::new(task.path.clone()).encode(),
-            task.identity.clone(),
-        ))
-        .expect("task fixture should serialize");
-        let record_bytes = SPILL_LENGTH_BYTES
-            .checked_add(u64::try_from(payload.len()).expect("task fixture length should fit"))
-            .expect("task fixture record size should fit");
-        let temporary_storage = TemporaryStorage::with_limit_bytes(
-            record_bytes
-                .checked_mul(2)
-                .expect("two task records should fit in the test limit"),
-        );
-        let (queue, _) = TaskQueue::new(root.clone(), 1, &temporary_storage)
-            .expect("scanner task queue should open");
-        queue
-            .schedule(task.clone())
-            .expect("first spill record should fit");
-        queue
-            .schedule(task.clone())
-            .expect("second spill record should fit");
-
-        let cancelled = AtomicBool::new(false);
-        let failed = AtomicBool::new(false);
-        let root_invalid = AtomicBool::new(false);
-        queue
-            .take(&cancelled, &failed, &root_invalid)
-            .expect("resident task should be available")
-            .expect("resident task should exist");
-        queue.complete();
-        queue
-            .take(&cancelled, &failed, &root_invalid)
-            .expect("first spilled task should be available")
-            .expect("first spilled task should exist");
-        queue.complete();
-        assert_eq!(temporary_storage.used(), record_bytes);
-
-        queue
-            .schedule(task.clone())
-            .expect("resident capacity should remain bounded");
-        queue
-            .schedule(task)
-            .expect("compaction should free capacity for the next spill record");
-        assert_eq!(
-            temporary_storage.used(),
-            record_bytes.checked_mul(2).expect("test limit should fit")
-        );
-        drop(queue);
-        assert_eq!(temporary_storage.used(), 0);
-    }
-
-    #[test]
     fn scanner_reports_temporary_storage_exhaustion_without_completing_the_root() {
         use crossbeam_channel::unbounded;
 
@@ -1547,42 +1097,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn task_queue_rejects_corrupt_spill_paths_outside_root() {
-        let root = PathBuf::from("/scan-root");
-        let (queue, _) = TaskQueue::new(root.clone(), 1, &TemporaryStorage::default())
-            .expect("scanner task queue should open");
-        let cancelled = AtomicBool::new(false);
-        let failed = AtomicBool::new(false);
-        let root_invalid = AtomicBool::new(false);
-
-        assert!(
-            queue
-                .take(&cancelled, &failed, &root_invalid)
-                .expect("initial scanner task should be available")
-                .is_some()
-        );
-        queue.complete();
-        {
-            let mut state = queue
-                .state
-                .lock()
-                .unwrap_or_else(std::sync::PoisonError::into_inner);
-            state
-                .spill
-                .push(DirectoryTask {
-                    path: PathBuf::from("/outside-root/secret"),
-                    identity: None,
-                })
-                .expect("corrupt spill fixture should be writable");
-            state.pending = 1;
-        }
-
-        let Err(error) = queue.take(&cancelled, &failed, &root_invalid) else {
-            panic!("out-of-root spill path should be rejected")
-        };
-        assert_eq!(error.kind(), io::ErrorKind::InvalidData);
-    }
     #[cfg(any(unix, windows))]
     fn replace_directory_with_link(path: &Path, displaced: &Path, target: &Path) {
         fs::rename(path, displaced).expect("original directory should be displaced");
